+++ CHANGELOG.md
@@ -10,9 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added 'importUrl' to metadata definition for easier driver updates. Just click Import on the driver and it'll grab the latest (But will overwrite any manual changes to the driver you have made, so be careful!).
+- Scheduled Run Time from Shark App (Preference) - Use this setting with Smart State Refresh Enabled to allow refreshes to lie dormant.  This setting should be set to the same time the Shark is scheduled to run through the Shark App.  If disabled, dormant refreshes default to 15 minutes. 
+- Schedule_Type (Attribute) - This Attribute shows the current type of refresh schedule your Shark device is following.
 
 ### Changed
 
+- State Detection Improvements - Fixed a bug where if Recharging to Resume was null, the robot would always show Returning to Dock.
+- Scheduling done by cron instead of RunIn - since cron scheduling survives a hub reboot, it is a more hands-off way of managing scheduling for your robot vacuum.  Once set, the schedule will faithfully run allowing any rules set up to auto run without intervention.
 - Updated 'status' state - Actually make it display one of four values: running, paused, returning to dock (previously undocked), or docked.
 
 ## [1.0.5]